requirements.txt
addict==2.4.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==26.1.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
chumpy==0.70
click==8.0.3
cloud-init==25.2
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
Cython==3.2.4
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.23.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isort==7.0.0
jeepney==0.7.1
Jinja2==3.1.6
json-tricks==3.17.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librt==0.8.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.7
mmpose==1.3.2
model-index==0.1.11
more-itertools==8.10.0
mpmath==1.3.0
munkres==1.1.4
mypy==1.19.1
mypy_extensions==1.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
onnx==1.20.1
onnx-simplifier==0.4.36
onnxruntime-gpu==1.17.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
ordered-set==4.1.0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pexpect==4.8.0
pillow==12.1.1
platformdirs==4.8.0
# Editable Git install with no remote (project1-pose-estimation==0.1.0)
-e /home/daniecraig_1099/project1-pose-estimation
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocotools==2.0.11
pycryptodome==3.23.0
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.3.2
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytokens==0.4.1
pytz==2022.1
PyYAML==6.0.3
requests==2.25.1
rich==14.3.2
scipy==1.15.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
service-identity==18.1.0
shapely==2.1.2
six==1.17.0
sos==4.9.2
ssh-import-id==5.11
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
termcolor==3.3.0
terminaltables==3.1.10
tomli==2.4.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.67.3
triton==2.1.0
Twisted==22.1.0
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
xkit==0.0.0
xtcocotools==1.14.3
yapf==0.43.0
zipp==1.0.0
zope.interface==5.4.0
